refactor: isolate Douyin media helpers
